formating and seeding
Removed server.js, as it was redundant with app.js. Added a dedicated README.md to the backend with descriptions of the backend's layout and the available terminal scripts. Added the "seed" script which runs the new file "userSeed.js" inside the Seeds folder. This populates the database's users collection with five profiles (can add more later) to make testing easy. Also added a .gitignore. Removed the hardcoded user addition in the app.js file, as this is done by the seeding file now.