Fixed seeding for user and listings, fixed backend and frontend running on same port
Updated the listing seeding and user seeding code to work in current framework; added script to run listing population; added script to completely clear the database; made the backend run on port 8001 to avoid issues arising from frontend and backend running on the same port, and updated code to account for this port change. Created dedicated folder  "images" in Seeds folder where the photos used to seed listings are stored, and finally updated the readme to account for these changes.